+++ lib/a_star_algorithm.dart
@@ -1,8 +1,4 @@
 library a_star_algorithm;
 
 export 'src/a_star_algorithm.dart';
-export 'src/connect_grid_ext.dart';
-export 'src/models/weight_point.dart';
-export 'src/create_grid_with_barriers_ext.dart';
-export 'src/create_grid_with_free_ext.dart';
-export 'src/find_steps_ext.dart';
+export 'src/weight_point.dart';
