--- lib/src/helpers/list_tile_ext.dart
@@ -1,8 +1,0 @@
-import 'dart:math';
-
-import 'package:a_star_algorithm/src/models/tile.dart';
-
-extension ListTileConverter on List<Tile> {
-  List<Point<int>> toPoints() =>
-      map((tile) => Point<int>(tile.position.x, tile.position.y)).toList();
-}
