+++ lib/src/weight_point.dart
